Fixed bias operations and simplified initializers
In some cases, a user might specify an adjacency matrix in
which some nodes in the layer are not connected to any input
nodes. In these cases, the output of the Sparse layer should
be zero for those nodes. Prior to this commit, this was not the
case: the bias vector did not comply with the adjacency matrix.
So when the bias was added, the aforementioned nodes would
incorrectly have some non-zero value added to them.

Now, we check if any of these 'disconnected' nodes are present
and warn the user if so. Additionally, we now fix the behavior
of bias terms: they now abide by the sparsity dictated by the
adjacency matrix. If a node is disconnected, it is now guaranteed
to output zero. Also corrected the count_params function to
account for this.

In addition, initialization has been greatly simplified and
improved. Instead of relying on some custom re-implementations
of builtin initializers, we use a simple lambda to ensure that
the initialized values comply with the adjacency matrix. Now,
the user can use any initializer they would like, and we don't
have to write mostly copy-pasted initializer code.